Adjust Grid to improve public API
Enumeration methods `#eachRow:` and `#eachCell:` are not needed as `#do:` can easily be used instead.

So far the only public API used for Grid is `Grid>>#eachCell` and `Grid class>>#of:by:` so these are the tested
methods. Testing of internal initialization methods is removed.

Note: testing by comparing collections has proven difficult as the differences in dynamically created `{}` and
compile time created arrays can cause confusion. So testing now focuses more on attributes such as the grid
having the correct number of Cell instances and having the correct number of rows.