Beginnings of an attempt to export mazes as a `pgm` image file
Initialiy I went down this route of exporting as pgm files due to their simplicity
and because at first glance the existing `ImageReadWriter` classes did not
allow for the low level pixel writing control I desired for drawing mazes as
`png` files.

I've stopped the `pgm` implementation here because I encountered 
difficulty in using the `ZnBufferedWriteStream` for writing the file.
I continually encountered errors when trying to put byte arrays on the steam.

I've found an alternative in the `Form` class. `Form` allows for the low
level pixel writing I need and can be used by the `ImageReadWriter` classes
as data to write as an image. Future development on this feature will
go down this route.