Implement Image writing using Form class for grid data
Implementing a `pgm` format image writer was too difficult as I couldn't
use `ZnBufferedWriter` as I intended. The subclasses of `ImageReadWriter` 
can use a `Form` instance as data. The `Form` class has the low level pixel
writing I need to create mazes as `png` files.

Implementation has begun using `MPMazeImage` as a utility class with no
instances created. However it looks like it will be a better idea for the class
methods to create and use an instance. This will be the next step from here.