Define public interface for `Grid` hiding cell implementation
The maze creating algorithms `MPBinaryTree` and `MPSidewinder` are tightly coupled to the
implementation of the cells in `Grid`. This makes it difficult to change the implementation of
the cells.

The changes expose a new interface for `Grid` that can be used by algorithms to create a maze.

The method `MPMazeAlgorithm class>>#on:` is renamed to `#applyTo:` to better suit the concept
of applying an algorithm to a grid.