Display ASCII representation of any `MPGrid` subclass
The `Grid>>#displayStringOn:` method has been appropriated to work accross the `MPGrid` hierarchy.
It is still implementation dependent so it is kept as an abstract method in `MPGrid`. Also an important
part of this commit is fixing an error in displaying instances of `MPBitfieldGrid`. The root of the bug
was how `#eachRows` was returning the collection. It was in the reverse order of what it needed to
be.

Although the displaying methods are on the instance they only use the public interface to create the 
ASCII representation of the grid. Three new methods have been made to be a part of this interface: 
`MPGrid>>#linkedEast:` and `MPGrid>>#linkedSouth`. The third is `MPGrid>>#eachRows` which
is needed because `#displayStringOn:` expects a two dimensional collection of the grid cells which
can be different from the internal grid implementation. 

I've considered making the ASCII display of the mazes an external class or trait in the future. I'm not
currently happy with the implementations of `#displayStringOn:` and have considered if stacks
would be useful instead of streams both internally and as data to work over. However for now it
works and I'm happy to move on an revisit it later.