Delegate maze string representation to a string building object
Building the string representation of a maze is a complex operation. Moving
the responsibility for building the string into a separate class will allow
easier refactoring, testing, and extension of the string building process.

This is the minium viable product change of creating a new class that works
and dosen't break printing the string representation. The existing tests
will need to be changed as they are likely testing for side effects that are now
less important in their new class.

All streaming helpers and constants related to building the string 
representation are moved to the `MPMazeAsString` class.

Cell tests are renamed with the project prefix.