Test end to end building of maze strings
The current streaming helper tests don't test for the end to end function of
`MPMazeAsString` instances and instead focus on the individual methods
of the `streaming helper` protocol. This is a legacy holdover from moving
the string building functionality from `MPGrid` instances. The intent is to
remove tests on these internal methods and just test the public interface
of the class.

Note: maybe `MPMazeAsString` should be renamed to `MPGridAsString`.
Grids can be used to make a maze but are not always a maze. This class
is only concerned with the string representation of a grid.