Refactor `MPMazeAsString` to use streams as instance variables
Previously methods in the `streaming -helpers` protocol used `WriteStream`
instances as arguments and returned their contents. This was  a legacy 
holdover from when building the string representation was a part of 
`MPGrid`. As these streams were used througout the code they are redefined
as instance variables. The matching tests are removed. The remaining tests
are focused on the public interface of `MPMazeAsString`.