Set grid coordinates to start with 0,0 at the top left corner
The coordinate system of the maze now begins with 0,0 at the top left of the
grid. Increases in the y axis move the point south down the grid while
decreases move the point north. This required inverting the concept of 
north and south throughout the code.

`MPGrid>>#eachRow` was adjusted so that each class would consistently
return a collection of rows where the first element of the first row would
be the cell at coordinate 0,0 in each implementation. This keeps the
method consistent enough to be an interface to the class.