Calculate and display the relative distances of cells in a grid
Distances are calculated using Dijkstra's algorithm.

`MPDistances` is a wrapper around a collection of associations between
cells and their distances from a given root. To display these distances the
new overridable `MPGrid>>#contentsOf:` method is created.

`MPDistanceGrid` is a concrete implementation that displays the distances
of its cells in the ASCII art output. This may be a temporary class. Any grid
implementation should be able to caculate and display the distances of
its cells. This may be better implemented as a decorator instead of 
subclassing.